% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truffle-package.R
\docType{package}
\name{truffle-package}
\alias{truffle}
\alias{truffle-package}
\title{truffle: Create Teaching Datasets with Known Effects Hidden in Messy Data}
\description{
Generate datasets with known statistical effects (truffles) that students
must rediscover, then bury them in realistic data-processing challenges
(dirt) that must be cleaned first. See \code{\link[=truffle_likert]{truffle_likert()}} for the main data
generator and the package README and vignette for worked examples:
\code{vignette("truffle", package = "truffle")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ianhussey/truffle}
  \item Report bugs at \url{https://github.com/ianhussey/truffle/issues}
}

}
\author{
\strong{Maintainer}: Ian Hussey \email{ian.hussey@unibe.ch} (\href{https://orcid.org/0000-0001-8906-7559}{ORCID})

}
\keyword{internal}
